Add P2P custom error
Having Box<dyn std::error:Error> for the prototype was great, however its making troubles as it makes it not safe for futures.

Its a CLI, initially we dont need to intercept this error at any point and react to it, so probably just giving an idea of the error to the end user is enough. It can scale as we find new requirements.