Fix termination due to timeout
We were sharing the same shutdown channel for every task. That caused other handshakes to be incomplete once the first of them ends. Lets keep it simple and locally. From here the upper level function will assume timeouts and user ctr+c terminations are going to be done on each subsystem. Thats not a problem, since all of them will be legitimate subscribers.