Do not break the program if one node fails
Before this change, if one of the nodes  had problems, that would affect the entire output of the program. With this change we avoid that, by assuming 2 kind of errors can happen:

* Unrecoverable program errors, not related to handshakes.
* Handshake related errors.

The former means an error in the runtime, so no other output should be provided to the user. The second, are expected errors and we will just add them to the new HandshakeResult struct.